Convert to TS - WIP
