+++ package.json
@@ -69,6 +69,7 @@
     "@babel/runtime": "^7.11.2",
     "@guardian/eslint-config-typescript": "^0.5.0",
     "@guardian/prettier": "^0.5.0",
+    "@types/doubleclick-gpt": "^2019111201.0.0",
     "@types/jest": "^26.0.20",
     "@types/lodash-es": "^4.17.4",
     "@typescript-eslint/eslint-plugin": "^4.12.0",
